+++ src/io/small_element_cache.cpp
@@ -27,7 +27,7 @@ bool SmallElementCacheStrategy::retrieve(rosbaz::io::byte* buffer, size_t offset
 
   assert(offset >= cache_found->offset);
 
-  std::copy_n(cache_found->data.begin() + static_cast<uint64_t>(offset - cache_found->offset), count, buffer);
+  std::copy_n(cache_found->data.begin() + (offset - cache_found->offset), count, buffer);
 
   return true;
 }
